Add radio button values to the context, independent of group membership
This is in preparation for adding support for JToggleButtons. Unlike radio
buttons, toggle buttons do not need to belong to ButtonGroups to be useful.
However, it's difficult to cleanly specify whether a button should be in the
context using overloads. Instead, I've decided that all button states will
be included in the context.